refactor(RegionsWithIntersections): move WithInterception outside of RegionsWithIntersections
