+++ packages/@sanity/base/src/presence/overlay/WithIntersection.tsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import {tap} from 'rxjs/operators'
+import {ObservableIntersectionObserver} from './intersectionObserver'
+
+interface WithIntersectionProps extends React.ComponentProps<'div'> {
+  onIntersection: (id, IntersectionObserverEntry) => void
+  io: ObservableIntersectionObserver
+  id: string
+}
+export const WithIntersection = (props: WithIntersectionProps) => {
+  const {onIntersection, io, id, ...rest} = props
+  const element = React.useRef<HTMLDivElement | null>(null)
+  React.useEffect(() => {
+    const el = element.current
+    if (!el) return undefined
+    const subscription = io
+      .observe(el)
+      .pipe(tap((entry) => onIntersection(id, entry)))
+      .subscribe()
+    return () => subscription.unsubscribe()
+  }, [io])
+
+  return <div ref={element} {...rest} />
+}
